chore: Extract components from BookingListItem.
