+++ apps/web/components/booking/RecurringBookingsTooltip.tsx
@@ -1,0 +1,87 @@
+import dayjs from "@calcom/dayjs";
+import classNames from "@calcom/lib/classNames";
+import { formatTime } from "@calcom/lib/date-fns";
+import { useLocale } from "@calcom/lib/hooks/useLocale";
+import { getEveryFreqFor } from "@calcom/lib/recurringStrings";
+import { BookingStatus } from "@calcom/prisma/enums";
+import { Tooltip } from "@calcom/ui";
+import { RefreshCcw } from "@calcom/ui/components/icon";
+
+import type { BookingItemProps } from "@components/booking/BookingListItem";
+
+interface RecurringBookingsTooltipProps {
+  booking: BookingItemProps;
+  recurringDates: Date[];
+  userTimeZone: string | undefined;
+  userTimeFormat: number | null | undefined;
+}
+
+const RecurringBookingsTooltip = ({
+  booking,
+  recurringDates,
+  userTimeZone,
+  userTimeFormat,
+}: RecurringBookingsTooltipProps) => {
+  const {
+    t,
+    i18n: { language },
+  } = useLocale();
+  const now = new Date();
+  const recurringCount = recurringDates.filter((recurringDate) => {
+    return (
+      recurringDate >= now &&
+      !booking.recurringInfo?.bookings[BookingStatus.CANCELLED]
+        .map((date) => date.toString())
+        .includes(recurringDate.toString())
+    );
+  }).length;
+
+  return (
+    (booking.recurringInfo &&
+      booking.eventType?.recurringEvent?.freq &&
+      (booking.listingStatus === "recurring" ||
+        booking.listingStatus === "unconfirmed" ||
+        booking.listingStatus === "cancelled") && (
+        <div className="underline decoration-gray-400 decoration-dashed underline-offset-2">
+          <div className="flex">
+            <Tooltip
+              content={recurringDates.map((aDate, key) => {
+                const pastOrCancelled =
+                  aDate < now ||
+                  booking.recurringInfo?.bookings[BookingStatus.CANCELLED]
+                    .map((date) => date.toString())
+                    .includes(aDate.toString());
+                return (
+                  <p key={key} className={classNames(pastOrCancelled && "line-through")}>
+                    {formatTime(aDate, userTimeFormat, userTimeZone)}
+                    {" - "}
+                    {dayjs(aDate).locale(language).format("D MMMM YYYY")}
+                  </p>
+                );
+              })}>
+              <div className="text-default">
+                <RefreshCcw
+                  strokeWidth="3"
+                  className="text-muted float-left mr-1 mt-1.5 inline-block h-3 w-3"
+                />
+                <p className="mt-1 pl-5 text-xs">
+                  {booking.status === BookingStatus.ACCEPTED
+                    ? `${t("event_remaining_other", {
+                        count: recurringCount,
+                      })}`
+                    : getEveryFreqFor({
+                        t,
+                        recurringEvent: booking.eventType.recurringEvent,
+                        recurringCount: booking.recurringInfo.count,
+                      })}
+                </p>
+              </div>
+            </Tooltip>
+          </div>
+        </div>
+      )) ||
+    null
+  );
+};
+
+export default RecurringBookingsTooltip;
